cardContainerBlock refactoring, remove cardBlock and make ImageEditor reusable
